rename to c2rust
