+++ src/c2rust/module-init.c
